Update the WPCS namespace
To prevent conflicts with WordPress itself - especially now that WordPress will let go of the PHP 5.2 minimum supported version -, WPCS 2.0 will not use the WordPress namespace.

While PHPCS 2.x still needed to be supported by WPCS, this could not be helped as it was a PHPCS requirement. Now that PHPCS 2.x support is being dropped in WPCS 2.0, the namespace will be adjusted to be `WordPressCS\WordPress`.

This PR updates our import statements and other class references to use the updated namespace.

Closes #299